improvement: add custom bart model for summarization of webpades and refactor code
