+++ solutionAgent/utilities/utils.py
